refactor: move QueryParams from oid4vci to shared utils module
