+++ crates/cloud-wallet-openid4vc/src/oid4vci/client/mod.rs
@@ -40,11 +40,11 @@ use crate::oid4vci::error::{
 };
 use crate::oid4vci::issuer_metadata::CredentialIssuerMetadata;
 use crate::oid4vci::notification::NotificationRequest;
-use crate::oid4vci::query_params::QueryParams;
 use crate::oid4vci::token_request::{
     AuthorizationCodeRequest, PreAuthorizedCodeRequest, TokenRequest,
 };
 use crate::oid4vci::token_response::TokenResponse;
+use crate::utils::QueryParams;
 use crate::utils::pkce::{derive_pkce_challenge, generate_pkce_verifier};
 
 type Result<T> = std::result::Result<T, ClientError>;
