+++ crates/cloud-wallet-openid4vc/src/oid4vci/mod.rs
@@ -15,6 +15,5 @@ pub mod css_color;
 pub mod error;
 pub mod issuer_metadata;
 pub mod notification;
-pub mod query_params;
 pub mod token_request;
 pub mod token_response;
